# Remove std::variant